refactor(massappct_u): Rename scraper
Change scraper from massappct128 to massappct_u
